fix(samples): make notebooks compatible with Qiskit 1.3
The notebook "Tomography of quantum teleportation" had the issue
"Invalid clbits for condition x" at transpilation with Qiskit 1.3 during
 HighLevelSynthesis with c_if() conditions.
To merge the two circuits of the notebooks, QuantumCircuit.append() was
used. But in this Github issue it's recommended to use
QuantumCircuit.compose() instead :
https://github.com/Qiskit/qiskit/issues/11379#issuecomment-1916517818
To make it work we had to change the size of the sub-circuit to be the
same at the other one (3 qubits and 3 classical bits).
Moving to QuantumCircuit.compose() fixed the error with the conditions.

Refs: #8785